refactor(#289): move config loader to root internal package
