+++ pkg/brc29/brc29_address_test.go
@@ -3,7 +3,7 @@ package brc29_test
 import (
 	"testing"
 
-	"github.com/4chain-ag/go-wallet-toolbox/pkg/internal/sdk/template/brc29"
+	"github.com/4chain-ag/go-wallet-toolbox/pkg/brc29"
 	ec "github.com/bsv-blockchain/go-sdk/primitives/ec"
 	sdk "github.com/bsv-blockchain/go-sdk/wallet"
 	"github.com/stretchr/testify/assert"
