+++ pkg/brc29/brc29_template_test.go
@@ -3,7 +3,7 @@ package brc29_test
 import (
 	"testing"
 
-	"github.com/4chain-ag/go-wallet-toolbox/pkg/internal/sdk/template/brc29"
+	"github.com/4chain-ag/go-wallet-toolbox/pkg/brc29"
 	"github.com/bsv-blockchain/go-sdk/chainhash"
 	ec "github.com/bsv-blockchain/go-sdk/primitives/ec"
 	"github.com/bsv-blockchain/go-sdk/script/interpreter"
